Add scripts for Zarr conversion using RasterIO and virtual environment setup; update sample creation scripts

+++ devtools/make_sample.zarr_big.sh
@@ -1,22 +1,31 @@
 #!/bin/bash
 
-set -xeuo pipefail
+set -euo pipefail
 
 cd "$(dirname "$0")"/../devdata
 
 TILE_NAMES=()
-for LAT in N{47..50}
+for LAT in N{47..49}
 do
         for LON in W00{0..3}
         do
                 TILE_NAMES+=("${LAT}${LON}")
         done
+        for LON in E00{0..1}
+        do
+                TILE_NAMES+=("${LAT}${LON}")
+        done        
 done
 
 OUT_NAME=FRnw
 
 for TILE_NAME in "${TILE_NAMES[@]}"
 do 
+  if [ -f "$TILE_NAME.zip" ]; then
+    echo "$TILE_NAME.zip already exists, skipping download."
+    continue
+  fi
+
   if curl https://openmaps.online/eudem_download/eu_4326/arc1/$TILE_NAME.zip -o "$TILE_NAME".zip
   then 
     rm -rf "$TILE_NAME"
@@ -27,19 +36,13 @@ do
 done
 
 rm -rf $OUT_NAME.zarr
-# Creation options (-co) precondition for our Micropython decoder (except BLOCKSIZE)
-# They are often default values, but we set them explicitly to avoid future changes
-gdal_merge.py \
-        -of Zarr \
-        -co "FORMAT=ZARR_V3" \
-        -co "CHUNK_MEMORY_LAYOUT=C" \
-        -co "DIM_SEPARATOR=/" \
-        -co "BLOCKSIZE=200,200" \
-        -co "COMPRESS=NONE" \
-        -o $OUT_NAME.zarr \
-        ./*/*.HGT
+rm -f $OUT_NAME.zarr.zip
 
-for TILE_NAME in "${TILE_NAMES[@]}"
-do
-  rm -rf "$TILE_NAME" "$TILE_NAME".zip
-done
+python ../devtools/convert_to_zarr.py --out $OUT_NAME.zarr ./*/*.HGT
+       
+
+
+# for TILE_NAME in "${TILE_NAMES[@]}"
+# do
+#   rm -rf "$TILE_NAME" "$TILE_NAME".zip
+# done
